Added WebVTT support and audio visualization
Lots of refactoring for efficiency and to start integrating the subtitle
timeline with the Ayamel framework. Audio still needs a lot of work for
processign efficiency and data acquisition, but the display is pretty
solid. Caption display is still separate from the timeline itself;
display and editing of captions is shown on the demo page.