Fixed up functionality for the different tools.
Added a button bar to the demo to switch tools, and then fixed all the
bugs that had crept in from not testing anything but the select tool. A
little bit of progress on improving asynchrony with the audio rendering,
and some progress towards tracking active tracks separately from the
current selection.