Switched timescale from ms to seconds.
The timeline now no longer requires scaling to interact with HTML5 media
elements that use seconds for time codes. This required some fiddling
with the key display, which is actually better now as a result, and the
slider, which was overdue for efficiency improvemets anyway. I tried
preloading images on Timeline.prototype to get rid of the problem with
the slider not rendering immediately, but that doesn't seem to have done
anything.
The ui was also updated to add cursor hotspots, include the first level
of track padding as part of the key, and auto-select newly created
segments.
The demo is updated with a button to add a new empty track.