Audio Resampling & Simplified Track Loading
Added a resampler to save memory in the waveform renderer. Speed doesn't
improve much- I already had that pretty well optimized with the display
sampling stride- but memory consumption is down by a factor of ~48 and
the display looks a lot cleaner.
Also moved the responsibility for loading track data via AJAX out of
Timeline.js and into the TimedText repository.